+++ tests/NoConstructorHydrator/Nested/Post.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Hydrator\Tests\NoConstructorHydrator\Nested;
+
+use LogicException;
+
+final class Post
+{
+    public string $title = '';
+    public ?Author $author = null;
+
+    public function __construct()
+    {
+        throw new LogicException('Constructor should not be called.');
+    }
+}